DRY linking a model to a report
