+++ django-backend/fecfiler/memo_text/models.py
@@ -1,12 +1,10 @@
 from fecfiler.soft_delete.models import SoftDeleteModel
 from fecfiler.committee_accounts.models import CommitteeOwnedModel
+from fecfiler.f3x_summaries.models import ReportMixin
 from django.db import models
 
 
-class MemoText(SoftDeleteModel, CommitteeOwnedModel):
-    report = models.ForeignKey(
-        "f3x_summaries.F3XSummary", on_delete=models.CASCADE, null=True, blank=True
-    )
+class MemoText(SoftDeleteModel, CommitteeOwnedModel, ReportMixin):
     rec_type = models.TextField(null=True, blank=True)
     filer_committee_id_number = models.TextField(null=True, blank=True)
     transaction_id_number = models.TextField(null=True, blank=True)
