+++ django-backend/fecfiler/scha_transactions/views.py
@@ -1,5 +1,6 @@
 from rest_framework import filters
 from fecfiler.committee_accounts.views import CommitteeOwnedViewSet
+from fecfiler.f3x_summaries.views import ReportViewMixin
 from .models import SchATransaction
 from django.db.models.query import QuerySet
 from .serializers import SchATransactionSerializer
@@ -7,7 +8,7 @@ from django.db.models import TextField, Value
 from django.db.models.functions import Concat, Coalesce
 
 
-class SchATransactionViewSet(CommitteeOwnedViewSet):
+class SchATransactionViewSet(CommitteeOwnedViewSet, ReportViewMixin):
     """
     This viewset automatically provides `list`, `create`, `retrieve`,
     `update` and `destroy` actions.
@@ -33,16 +34,6 @@ class SchATransactionViewSet(CommitteeOwnedViewSet):
         .all()
     )
     """QuerySet: all schedule a transactions with an aditional contributor_name field"""
-
-    def get_queryset(self):
-        report_id = None
-        if self.request is not None:
-            report_id = self.request.query_params.get("report_id")
-
-        queryset = super().get_queryset()
-        if isinstance(queryset, QuerySet):
-            queryset = SchATransaction.objects.all().filter(report_id=report_id)
-        return queryset
 
     serializer_class = SchATransactionSerializer
     permission_classes = []
